Remove un-needed interface
